devicemanager: known device cache
Keep track of known device and dump their info to cache. This should
speed up reconnect after restarting mconnect. The rationale is that the
devices are likely to be connected to to the local network and use
DHCP and as such will most probably get the same address the next time
they appear. Hence, it should be ok to attemp connection the last know
IP address. If that fails, the discovery process will still take place.