clipboard: Add incoming clipboard support
Add support for incoming kdeconnect.clipboard messages. When you copy text on
a paired device, it will be pasted to the clipboard on the local device.

Note that this is separate from the text sharing feature. Most notably, I've
decided not to display a notification when the clipboard is synced, since that
could get annoying.